requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.1
basicsr==1.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dill==0.3.6
efficientnet_pytorch==0.7.1
einops==0.8.0
facexlib==0.3.0
filelock==3.14.0
filterpy==1.4.5
fonttools==4.52.4
fsspec==2024.5.0
future==1.0.0
gfpgan==1.3.8
grad-cam==1.4.8
grpcio==1.64.1
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.1.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
sympy==1.12
tb-nightly==2.18.0a20240616
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.3
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.4
transformers==4.41.2
triton==2.1.0
ttach==0.0.3
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
yapf==0.40.2
zipp==3.19.2
